samfair/requirements.txt
numpy>=1.26.4
scipy>=1.14.1
openai>=1.0.0
pinecone-client>=5.0.0
python-dotenv>=1.0.1
pymupdf>=1.24.10
pytesseract>=0.3.13
Pillow>=10.4.0
rich>=13.9.1